pyproject.toml
[project]
name = "sih-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "alembic>=1.16.5",
    "bcrypt>=4.3.0",
    "fastapi[standard]>=0.116.1",
    "geoalchemy2>=0.18.0",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.10.1",
    "python-jose>=3.5.0",
    "shapely>=2.1.1",
    "sqlalchemy>=2.0.43",
    "sqlmodel>=0.0.24",
    "web3>=7.13.0",
]
